Split metrics module into multiple files
This commit is the start of batching work in metrics. We know we pay a high
price for our metrics, so let's hide that cost.
